requirements.txt
langchain
langchain-groq
langchain-community
langchain-core
python-telegram-bot
streamlit
supabase
python-dotenv
tavily-python
pandas

